feat: pix2pix task for palette

+++ scripts/run_tests.sh
@@ -223,6 +223,23 @@ if [ $OUT != 0 ]; then
     exit 1
 fi
 
+####### pix2pix test
+echo "Running pix2pix tests"
+URL=https://joligen.com/datasets/SEN2VEN_mini.zip
+ZIP_FILE=$DIR/SEN2VEN_mini.zip
+TARGET_PIX2PIX_DIR=$DIR/SEN2VEN_mini
+wget -N $URL -O $ZIP_FILE
+mkdir $TARGET_PIX2PIX_DIR
+unzip $ZIP_FILE -d $DIR
+rm $ZIP_FILE
+
+python3 -m pytest -p no:cacheprovider -s "${current_dir}/../tests/test_run_pix2pix_diffusion.py" --dataroot "$TARGET_PIX2PIX_DIR"
+OUT=$?
+
+if [ $OUT != 0 ]; then
+    exit 1
+fi
+
 echo "Deleting target dir $DIR"
 rm -rf $DIR/*
